WIP Merging scenes for 59icosahedra
This takes a different approach to toggling cell orbits.  Rather than scripting macros to operate
on a single vZome design, we have several predefined models, one per cell orbit, and combine
the scenes in a single canvas.

To accomplish this, I had to decouple the CameraStateContext from the WorkerContext.

STATUS:
- single cell orbit toggling is working great
- background click is not working at all
- background color for the stellation canvas needs to be controlled
- initial camera also