WIP Supporting vzome-legacy.js module in quick workflow
It should be possible to load the published module and run `classic` or `59icosahedra` in the
quick workflow.  The script changes are working fine, and the vZome files can be loaded and
interpreted correctly... but somehow further actions are not working because we call a module
where `designWrapper` is undefined.