WIP Responsive layout
I moved most of the styling out of `index.jsx` and into `responsive.css`.

I'm taking a "mobile first" approach, and the layout for phones is pretty good.
However, it is only avoiding "canvas runaway" due to a `max-height` in the
`selector` class.  I rearranged the sections in `below-appbar` to make this work.

I have a media query for larger screens commented out at the moment.