No more worker use for toggling cell states
I'm managing the instance `selected` state for labeled panels directly, with no
round-trip to the worker and controllers.

I'll be able to use the preview `.shapes.json` files for the two selector designs,
once I make sure the labels from the groups in the `.vZome` file (manually added)
are manually added to the preview file.  This will make the startup time much faster,
with no use of `vzome-legacy.js` at all.