Added distance prop to CameraStateProvider
I moved `initialState` from `workerClient/actions.js` to `camera.jsx`, and
refactored it.

Distance prop is working nicely on all three cameras.  The same value of 500 works well
in both cases, it turns out.

SelectorCanvas is subscribed to the worker client messages, to get the camera position
and orientation from the selector models, and set it on the cameras.

I also copied the camera state from `pieces-bdfh` to `pieces-aceg`, so they are in sync
from the start.