Removing circular module dependency
The `vzome-legacy` module was importing `getSceneIndex` from the static worker
module, which is also the module that imports `vzome-legacy` dynamically.  This
is harmless in normal situations, but disastrous in the "quick" workflow, which loads
the dynamic module from `vzome.com`.  It results in very odd behavior, since two
copies of the same module are present at runtime.

This change is the most obvious, simple way to fix the dependency, but needing the
`getSceneIndex` function in both places does point at a leaking abstraction boundary.