WIP Complete refactoring of camera state and trackball
I'm taking a largely new approach, but one that is very rational and clear.  The key
idea is that the `TrackballControls` do not manipulate the actual Three.js camera,
but instead they operate on the `trackballCamera` managed by the `CameraContext`.
The context also subscribes to the `change` event from the trackball; when the event
fires, the context-managed camera state (vZome-like) is updated to match the
`trackballCamera`.

The other half of this strategy is that the `PerspectiveCamera` is configured by a
context-managed `perspectiveCamera` store, which is derived from the main camera
state.

This gives ownership of the camera completely to the context, which is essentially
interposed between the `TrackballControls` and the `PerspectiveCamera`.
It therefore makes it trivial to share camera state across two canvases and their
`PerspectiveCamera`s.

The second `CameraProvider`, for `CameraControls`, is nested in the first.
For a moment, I had an extra layer so that the `useCamera` and the
`CameraProvider` were not both in the same component, since I thought that was
causing a defect.  The actual defect is described on line 121 of the `CameraProvider`.

When the `CameraProvider` has an injected context, it behaves slightly differently.
First, it sets its rotation state from context.  Second, its trackball sync updates the
context camera with rotation, not the local one, and ignores pan+zoom.  Finally,
only pan+zoom is accepted from loaded trackball designs, for local state.

STATUS:  Everything is working, including the contextual menu camera commands.