Support show-scenes="named"
Now the intended value space for `show-scenes` on the web component is:
"all", "named", or "none".  You can still use "true" (==="all") and "false" (==="none").
If the value is "named", then only explicitly named scenes will be shown.  This also
implies that the default scene will not be shown.

I also fixed the bug involving export resources loading.  Now the event is only triggered
for classic.