Zomic support restored for Online classic
This lets me have a completely different implementation for Javascript.
I have added RunZomicScript to the JSweet transpile, and it now initializes the
module, and injects it into the `JsAlgebraicField`.

I don't have a way to move the Zomic support classes from `core` into the new
`vzome-zomic` module yet.  The new module is about 250K, mostly ANTLR runtime.
The Zomic classes added about 0.1MB to `core-java.js`, which was 2.7MB.

The "compiler" is working!  It took a bit of debugging and guesswork, since the `ctx`
API is not well documented for Javascript.

I had to make a couple of adjustments to `ZomicCompilerState`
to get it through JSweet, including removal of some bad overloading.