Startup macros
I have improved the prior work on custom menu macros, and I can now
support exports and captures, since the `ControllerFileAction` will be in
`headless` mode when used in a macro.

Furthermore, this now supports a command-line property, `-macro`, whose
value must be a comma-separated list of command strings.
I've captured documentation in the workflow doc.