Using lightweight GLJPanel to avoid spin
On my Mac OS Sonoma, vZome locks up if I use Finder to open a file.
For Ronny Bagdadi, *any* gesture to open a file causes lock-up.
I've switched from heavyweight GLCanvas to lightweight GLJPanel, and
the problem is gone.  The frame rate is much worse, but I can live with
that.  There may be configuration I can do to improve it.