WIP GitHub sharing from Online
This branch takes a much simpler approach!  The user must generate and
provide a Personal Access Token.  This lets us simply bypass the whole
OAuth flow!

STATUS:

The feature is working, with a nice explanatory dialog to enter the PAT,
which gets placed in local storage for the next use. The dialog shows an
"uploading..." state, and the "Share" button is disabled appropriately.

It successfully commits a complete vZome file to the root of my repo,
including README.md, index.md, preview JSON, and preview image,
then opens a new tab with the share URL.

The commit is now happening from the worker.

TODO:
- remove org/repo/branch hardcoding
- allow config
- persist config
- test more failure modes