Completed config back-end for Online sharing
I don't have a sharing config UI yet, but everything else is ready for it.

I also decided to move the `.vZome` XML serialization into the worker.
This is mostly working, using `JavaDom*` instead of client DOM, except
that scenes from an open document are lost!  Also, the attributes should
be sorted before while serializing them.

I'm now validating the PAT before storing or using it, and storing the org name
(the username, for now), repo name, and branch name together with the token.
I have added the ability to choose another org or repo.

Later, I'll add the option to create a repo.