Added config screen for online sharing
I have implemented all fields in the config screen.
The style is defaulted appropriately based on existing scenes.

I fixed an issue with loading the camera from XML.

I also fixed a *desktop* issue where the title and description were ignored, and
I enabled a style value of "menu (all)" in GitHubShare.java.

All the "config" data is now stored in the editor state, so it can differ per document,
but persists *in session* for multiple shares.  Eventually I suppose I'll persist it
with the file.