Combined selection & strut drag tool
I have refined the event handling in `interaction.jsx` so that clicks and drags are
mutually exclusive.  A drag is started only when sufficient movement has happened,
and a click is generated when not dragging and not too much time has passed.

The UnifiedTool is a merge of StrutDragTool and SelectionTool.  I should change
it so it reuses those two instead, but that is too much change today.