Added "reset orbit colors" command
This allows one to open a design with custom (or outdated) strut colors
and reset the orbit colors in the SymmetrySystem.

Note that "reset colors" on a selection is still a separate gesture.

I also fixed the defect for ball colors, wherein Color.WHITE was used instead
of the color preference.

Finally, in preparation for further work on this branch, the SVG files for
the parts list in online now have their fill colors stated just once each, so
that I'll be able to programmatically color the images in the table, once
I am handling colors correctly in the BOM.