Porting POV-Ray and other exporters
I did the necessary Java refactoring to allow `PovRayExporter.java` and some
others to transpile with JSweet.  This involved introducing `DocumentIntf` and
`CameraIntf`.

I did some of the work on the Javascript side, but I haven't completed the
key `createDocument` function to create the DocumentIntf instance yet.

Along the way, I noticed that PartGeomExporter probably does not work on
the Java side, since its `exportDocument()` method won't be called.