WIP Camera tweening
Zoom, rotation, and pan tweening are all working, though I don't know how to
interrupt it on mouse interaction yet.  Nonetheless, it seems to work nicely even
when interrupted rapidly.

I need to give more control in both the APIs (web component and Solid) and for
the user in settings for the web component.