Added short strut mechanism
ExportedVEFStrutGeometry can now have a fallback for when the geometry
does not work.  If any non-tip vertex computed is further from the origin than
some tip vertex, the fallback geometry is tried, or null is returned if there is no
fallback.

I'm using FastDefaultStrutGeometry as a default fallback, so there
should always be a strut rendered; this should retain invariants from before.
In particular, switching styles will do the right thing as long as every strut
is rendered; if a strut does not get rendered, it cannot be recovered by
switching styles.

This solves Paul's issue with "inside-out" r00 struts, but it goes much further,
and avoids rendering *any* inside-out struts except for the very short default
struts.