Reviving old inspector UI with current worker
I've copied the minimal set of r114 modules required to run the inspector client, but
with two changes to `chunk-QXUNN3TL.js`:
 - it creates a worker from `/modules/vzome-worker-static.js`
 - it stamps every message to the worker with `isInspector: true`
This will let me match the old scene protocol responses, and thus I can debug
vZome files using the latest online code.

I have started the support for the older scenes, by passing `polygons` into
`scenes.js` `realizeShape()`.  The rendering is now working fine.  I could perhaps find
a cleaner way to pass in `polygons`.

I have done some refactoring, and the inspector and debugger now work.
However, there is still the original defect I have been trying to isolate, the reason I was
restoring this debugger: undoing a branch does not properly undo all the edits.
This is an interpreter defect in the worker, so it doesn't matter whether the UI is
modern code or the old inspector, the behavior is the same.

The worker sends back the last edit executed, but the inspector is currently ignoring
that; it should be setting the selection after.