New design and open URL now mutually exclusive
This might resolve the undefined `__store` defect, since the controller
hierarchy in the client context is now much more predictable.

I also improved the `ErrorBoundary` behavior for classic online, to log
the error as well as display it.  I could do better here, with more information
for the user, like how to report the error.