Various layout improvements for online
This correctly scales with window size, to a point.  There will be an issue if there are too
many selection bookmarks.

I opted to make the build/parts/measure tabs use a scroller, to keep things simple
for very short (not much height) windows.

The toolbar scrollbars look awful.  I'll have to fix it eventually.