WIP Converting 4OFF (Stella4D) file to 4D mesh JSON
I haven't nailed down the JSON format yet, so for now it is just an object in
the console.  I should probably use the format that Nan and I agreed upon
for the Observable notebook on Exploding 4D Polytopes.

Note that there are no edges being computed yet; the OFF format never has edges.