Restored Zomod support to desktop vZome
I also added Zomod support to the PrintVisitor.  Note that the PrintVisitor does
NOT output valid Zomic, even if the program was parsed as Zomic.

I had to modify `ZomicEditorPanel.readTextFromFile()` to do conversion from
CRLF (Windows, just in case) and CR (Mac, most of the files) to LF line endings,
since the JTextArea ignores CR as line endings.

I added a debug launch for desktop vZome.  It is missing a lot of JVM config, but
it functions mostly.