Add support for Hedron OFF
I think this format (with edges as 2-arity "faces") is due to Jim McNeill's Hedron
app, but I'm not certain.

The code should still work for the 4OFF format, though I haven't tested that after
my changes here.

The `field-api` code is now adjusted to generate values for the algebraic vertex data
of the compound of 20 octahedra.  The real values won't match those in the `values`
table in the `convert-40ff` test, since those have been scaled to match the actual
OFF values I have.