Attempting to detect empty image captures
Both for sharing and normal image save, an error should be reported if the
blob comes up empty.  If we see zero-length images still happening in GitHub,
without any such errors appearing,
then we know that the error is in the REST API use, not in producing the image.

I also added logging for each image capture, right at the canvas level.  Note that
the r3f discussion linked from the code indicated we should be setting
`preserveDrawingBuffer` on the GL context, and I don't see that we are.
I'm not sure how to do that through solid-three, yet.