Add MATHML_FORMAT and MATH_FORMAT to Java to support transpiled JS code
In the Java code, MATH_FORMAT produces the same result as DEFAULT_FORMAT, and is only used in the test cases to be sure no overflows are generated.

At some point, it could actually be implemented in Java to match the corresponding JavaScript format.

Adding this fix to the Java build avoids overwriting the fix in core-jave.js from commit 671cab114fc5ab6f413ca1d50b3bc8f753c933b3 when jSweet is run.