Component context API improvements
I have removed `#cameraStore` from the `vzome-viewer`, `vrml-viewer`,
and `zometool-instructions`, and everything is working fine.

The `CameraContext` now supports `tweening` and `lighting` properties
for initialization, and a `setTweenDuration` API method.

I'm trying to adopt a general mechanism for accessing context APIs,
perhaps exposing `useCamera` in the `ViewerContext`, etc., but I
have not taken that step yet.