WIP Viewer preview in sharing dialog
I have rearranged the dialog order, so the config page (now with a
viewer preview) comes first.  I also moved the blog and publish
checkboxes to the page where the repo choice is made, since those
are all GitHub specific, and won't apply when I add "save to folder".

I have enabled the "download" icon in all preview styles, but disabled the
"open in online" link when in the editor.

Layout for the left side of `ConfigPage` is iffy.

I ended up doing a significant refactor of the client code:
- removed the scene state from`ViewerContext`
- renamed requestSnapshot as showIndexedScene
- restored load-camera feature for next/prev
- simpler callbacks API for web component context
- factored SceneViewer out of DesignViewer
- factored all behaviors out of `SceneProvider`
- scene style normalization in the right two places
- new scene style "given" internally, and "named" is now "titled"
- new scene styles "indexed"
- added `SceneTitlesProvider` and `InitializeScene`

STATUS:
- online sharing preview slight defects:
   - mode switch does not reinit title
   - poor layout for steps buttons, using overflow