Better error handling in online vZome
The debugger now reports more useful errors, and loading a design
now fails if the old "bug accommodation" mechanism is hit.

I'm also detecting no-op edits while interpreting a design.
Old files have this a lot, so I cannot make this a failure!

I have revived the `regression.html` file.  I had to make a
fix to get the "success" events to generate correctly.

I have also added regression capability to the `localfiles` app,
behind a query parameter.  I had five failures out of the 114
files in the core regression suite, but I have already fixed the
one about `sqrt2` in vZome 2.0 files, and apparently that fixed
two of the five.

One of the three remaining failures is not meant to be tested...
it requires `CommandSparseCrossProduct`, which does not
exist any more!  The Java regression skips this design.  I have
renamed the file so it gets skipped for online also.

I also fixed a minor bug: the error alert was not cleared when a
new design is loaded.