WIP Added batch processing to localfiles app
I've added `exportAs()` to the `ViewerContext`, and created a
dedicated `EXPORT_TRIGGERED` message to support that, so we
don't need the controller and the `EditorContext`.  The `exportAs`
function becomes part of the API exposed to the web component.

I had to fix the viewer API for the app to work at all and render
the files, before I even started the export work.  I also had to
incorporate the `lastSceneIndex` bug fix from another branch.

The UI works, including writing the files using the granted permission
for the directory.  Typical exporters work, but error reporting is not
great (try POV).

I still need to add image capture, and set MIME types correctly.

Note that a zero-length file is created as soon as the file handle is
created, well before we write to it.