I corrected the file extension for POV-Ray exports.
I moved `indexResources()` to the `ViewerContext`, and added
a call in the `WebComponentBindings`.

I added a design test in the worker, to prevent an attempt to do
most exports on a preview.

I re-enabled all the POV and VRML elements, and added a
placeholder for glTF.

Finally, I corrected the layout for the `localfiles` app.